+++ packages/backend/src/models/CatalogModel/CatalogModel.ts
@@ -11,7 +11,6 @@ import {
     type ApiCatalogSearch,
     type ApiSort,
     type CatalogFieldMap,
-    type CatalogFieldWhere,
     type CatalogItem,
     type CatalogItemSummary,
     type CatalogItemWithTagUuids,
@@ -210,6 +209,7 @@ export class CatalogModel {
             .column(
                 `${CatalogTableName}.catalog_search_uuid`,
                 `${CatalogTableName}.name`,
+                `${CatalogTableName}.label`,
                 'description',
                 'type',
                 `${CachedExploreTableName}.explore`,
@@ -857,16 +857,24 @@ export class CatalogModel {
             })
             .innerJoin(
                 { source_metric: CatalogTableName },
-                `${MetricsTreeEdgesTableName}.source_metric_catalog_search_uuid`,
-                `source_metric.catalog_search_uuid`,
+                function joinSource() {
+                    void this.on(
+                        `${MetricsTreeEdgesTableName}.source_metric_catalog_search_uuid`,
+                        '=',
+                        `source_metric.catalog_search_uuid`,
+                    ).andOnVal('source_metric.project_uuid', '=', projectUuid);
+                },
             )
             .innerJoin(
                 { target_metric: CatalogTableName },
-                `${MetricsTreeEdgesTableName}.target_metric_catalog_search_uuid`,
-                `target_metric.catalog_search_uuid`,
-            )
-            .where('source_metric.project_uuid', projectUuid)
-            .andWhere('target_metric.project_uuid', projectUuid);
+                function joinTarget() {
+                    void this.on(
+                        `${MetricsTreeEdgesTableName}.target_metric_catalog_search_uuid`,
+                        '=',
+                        `target_metric.catalog_search_uuid`,
+                    ).andOnVal('target_metric.project_uuid', '=', projectUuid);
+                },
+            );
 
         return edges.map((e) => ({
             source: {